DEBUG variable, graceful exits, final packet clip.
Clip off the final packet so that it is the correct amount of bytes
and not just 500 with extra "0" bytes.

Created a new DEBUG variable that controls the output of debug-type
information onto the console (keep it on during development, turn it
off once turned in, etc).

Cleaned up code a little bit

Added some code to make more graceful exits ( incorrect arguments,
deal with divide by zero, etc.)